server/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
ultralytics==8.3.40
python-multipart==0.0.9
Pillow==11.0.0
